// src/interfaces/TelemetrySafetiSapi.ts
export interface TelemetrySafetiSapi {
  MAND: string;
  VWERK: string;
  NUM_ECON: string;
  ACEL_G_MAX: number;
  ACEL_G_PROM: number;
  LATITUD: number;
  LONGITUD: number;
  DIA: string;
  HORA: string;
  FRENO_G_MAX: number;
  FRENO_G_PROM: number;
  GIRO_PROM: number;
  GIRO_MAX: number;
  ID_PROV: string;
}